backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary>=2.9.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.8.2
